requirements.txt
pandas
optax
dill
dm-haiku
jax[tpu] -f https://storage.googleapis.com/jax-releases/libtpu_releases.html
jaxline
numpy
tensorflow
tensorflow-datasets
google-cloud-storage
wandb
dm-tree
natsort
fabric
rich
etils[eapp,edc,etree]
kaggle
gcsfs